requirements.txt
Django==4.2
django-environ==0.11.2
django-filter==23.2
djangorestframework==3.14.0
mysqlclient==2.2.0
django-cors-headers==4.2.0

